android: bypass local dns servers
