+++ nym-vpn-core/crates/nym-firewall-config/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "nym-firewall-config"
+version.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+documentation.workspace = true
+edition.workspace = true
+license.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+ipnetwork.workspace = true
+
+[lints]
+workspace = true
