+++ nym-vpn-core/crates/nym-vpn-lib/Cargo.toml
@@ -109,12 +109,14 @@ nym-windows.workspace = true
 nym-routing.workspace = true
 nym-dns.workspace = true
 nym-firewall.workspace = true
+nym-firewall-config.workspace = true
 
 [target.'cfg(any(target_os = "linux", target_os = "macos"))'.dependencies]
 hickory-server = { workspace = true, features = ["resolver"] }
 nym-routing.workspace = true
 nym-dns.workspace = true
 nym-firewall.workspace = true
+nym-firewall-config.workspace = true
 
 [target.'cfg(target_os = "android")'.dependencies]
 android_logger.workspace = true
